* Moved spectroscopy to characterization.
