+++ qiskit_experiments/characterization/spectroscopy.py
